requirements.txt
django
django-ninja
gunicorn
django-heroku
dj-database-url
